Moved your location movement into trufi map controller, renaming
