+++ lib/plan/plan_empty.dart
@@ -2,8 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong/latlong.dart';
 
-import 'package:trufi_app/blocs/location_provider_bloc.dart';
-import 'package:trufi_app/widgets/alerts.dart';
 import 'package:trufi_app/widgets/trufi_map.dart';
 
 class PlanEmptyPage extends StatefulWidget {
@@ -43,21 +41,12 @@ class PlanEmptyPageState extends State<PlanEmptyPage> {
     return FloatingActionButton(
       backgroundColor: theme.primaryColor,
       child: Icon(Icons.my_location, color: theme.primaryIconTheme.color),
-      onPressed: _handleOnMyLocationTap,
+      onPressed: _handleOnYourLocationPressed,
       heroTag: null,
     );
   }
 
-  void _handleOnMyLocationTap() async {
-    final locationProviderBloc = LocationProviderBloc.of(context);
-    LatLng lastLocation = await locationProviderBloc.lastLocation;
-    if (lastLocation != null) {
-      _trufiOnAndOfflineMapController.mapController.move(lastLocation, 17.0);
-      return;
-    }
-    showDialog(
-      context: context,
-      builder: (context) => buildAlertLocationServicesDenied(context),
-    );
+  void _handleOnYourLocationPressed() async {
+    _trufiOnAndOfflineMapController.moveToYourLocation(context);
   }
 }
